APPS-20 APPS-21 APPS-25 6.0 support (#75)
* feat: APPS-20 Added batch write, delete, udfs operations to existing batch read.

Old batch read api is deprecated. The new api accepts arbitrary combinations of batch objects in the body.

BREAKING CHANGE

* feat: APPS-25 Support partition query with pagination. 

The scan/ endpoint is now DEPRECATED in favor of the more performance query/ endpoint

* feat: Allow for multiple CTX objects to be provided for map and list operations.

The old operation model only allowed a user to nest into a single level of a CDT.  Now you will be able to provide an arbitrary number of CTX objects.

BREAKING CHANGE

* dependencies: Update java client to 6.1.2

This add support for partition queries and batch write, delete, and udfs.  This also removes support for the DEPRECATED predicate expressions and map decrement operation.

BREAKING CHANGE

* feat: Accept GeoJSON bin values mime type `application/json`.

 GeoJSON objects are still not supported within CDTs with this mime type.  The old method of base64 encoding the GeoJSON object is now DEPRECATED